refactor v2 eval string operators
